// src/types/todos.types.ts
export type Todo = {
  id: string; // UUID generated by Supabase
  title: string;
  content: string;
  done: boolean; // default FALSE
  created_at: string; // Auto-generated timestamp
  updated_at: string | null; // Auto-generated timestamp on update
};

export type NewTodo = Omit<Todo, 'created_at' | 'updated_at' | 'id' | 'done'>;

export interface TodosResult {
  data: Todo[];
  next_page: number | null;
  previous_page: number | null;
  total_pages: number;
}